requirements.txt
Django>=5.2.1
django-widget-tweaks>=1.5.0
djangorestframework>=3.14.0
Pillow>=10.0.0